Emitter thread pool opt. and message unique context
