+++ messaging/messaging/src/test/java/io/helidon/messaging/LatchedTestData.java
@@ -15,6 +15,8 @@
  *
  */
 
+package io.helidon.messaging;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
